# src/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Arbesk API",
    "description": "Cloud-Native 4D Fractal Version-Controlled 3D Asset Platform \u2014 Backend API for asset generation, session management, IPFS upload credentials, burn cleanup, and Nostr comment archival. Manifest writes, parametric editing, manifest-chain walks, and token resolution are handled client-side.",
    "version": "1.0.0",
    "contact": {
      "name": "Arbesk",
      "url": "https://github.com/ahmadsayed/arbesk"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:9090/api/v1",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/ipfs/upload-url": {
      "post": {
        "tags": ["IPFS"],
        "summary": "Mint a short-lived client upload credential",
        "description": "Session-gated and rate-limited per wallet. In Pinata mode returns a presigned upload URL; in Kubo mode returns the local API URL. The master Pinata JWT never reaches the client.",
        "operationId": "mintUploadCredential",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Upload credential returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "backend": {
                      "type": "string",
                      "enum": ["kubo", "pinata"]
                    },
                    "url": {
                      "type": "string",
                      "description": "Presigned upload URL or Kubo API endpoint"
                    },
                    "expiresAt": {
                      "type": "integer",
                      "description": "Unix milliseconds expiration time (Pinata only)"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
    "/ipfs/unpin": {
      "post": {
        "tags": ["IPFS"],
        "summary": "Unpin all IPFS CIDs owned by a manifest chain",
        "description": "Called before token burn or after asset removal from a collection (the backend verifies ownership on-chain, so the token must still be live). The session wallet must own the token or be an editor (Merkle proof), and `cid` must belong to the claimed token: accepted when it equals the tokenURI CID or appears in the `assets` map of the collection manifest at tokenURI or of up to 5 `prev_asset_manifest_cid` ancestors. A body-supplied `contractAddress` must be one of the chain's configured contracts (free or paid tier); when omitted, both are tried in order. Walks `prev_asset_manifest_cid` backward, collecting manifest CIDs, thumbnail CIDs, and comments archive CIDs, then unpins them so they become GC-eligible.",
        "operationId": "unpinManifestChain",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["cid", "tokenId"],
                "properties": {
                  "cid": {
                    "type": "string",
                    "description": "Starting manifest CID (typically the latest version)"
                  },
                  "tokenId": {
                    "type": "string",
                    "pattern": "^[0-9]+$",
                    "description": "Collection token ID the CID belongs to (decimal string, uint256-safe)"
                  },
                  "chainId": {
                    "type": "integer",
                    "description": "Chain ID (defaults to the backend's configured chain)"
                  },
                  "contractAddress": {
                    "type": "string",
                    "pattern": "^0x[a-fA-F0-9]{40}$",
                    "description": "Optional contract override; must be one of the chain's configured contracts (free or paid tier)"
                  },
                  "proof": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "pattern": "^0x[a-fA-F0-9]{64}$"
                    },
                    "description": "Merkle editor proof (bytes32 hex entries); required for non-owners"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unpin completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "unpinned": {
                      "type": "array",
                      "items": { "type": "string" },
                      "description": "CIDs that were unpinned"
                    },
                    "errors": {
                      "type": "array",
                      "items": { "type": "string" },
                      "description": "Per-CID unpin errors (do not fail the whole request)"
                    },
                    "durationMs": {
                      "type": "integer",
                      "description": "Time spent walking and unpinning"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (VALIDATION_ERROR), invalid or unknown token (INVALID_TOKEN), contractAddress not in the chain's configured contracts (INVALID_CONTRACT), or CID not referenced by the token (CID_NOT_IN_TOKEN)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "401": {
            "description": "Missing or invalid session",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "403": {
            "description": "Session wallet is not the token owner or an editor (FORBIDDEN)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          },
          "502": {
            "description": "Token's collection manifest could not be read (COLLECTION_UNREADABLE)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Session authentication. Format: `Session <opaque-token>`. Obtain a token by signing a SIWE (EIP-4361) message via `POST /api/v1/sessions`. Tokens expire after 24 hours."
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message"],
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code",
                "example": "REPLAY_DETECTED"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message",
                "example": "This transaction has already been consumed"
              },
              "details": {
                "description": "Optional additional error context",
                "oneOf": [
                  {
                    "type": "object"
                  },
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            }
          }
        }
      }
    }
  }
}